:truck: move `CertificationPointOfContact` to specific context
