+++ api/src/identity-access-management/domain/read-models/CertificationPointOfContact.js
