+++ api/src/identity-access-management/infrastructure/repositories/certification-point-of-contact.repository.js
@@ -2,8 +2,8 @@ import _ from 'lodash';
 
 import { knex } from '../../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../../shared/domain/errors.js';
-import { CertificationPointOfContact } from '../../../shared/domain/read-models/CertificationPointOfContact.js';
 import { AllowedCertificationCenterAccess } from '../../domain/read-models/AllowedCertificationCenterAccess.js';
+import { CertificationPointOfContact } from '../../domain/read-models/CertificationPointOfContact.js';
 
 const CERTIFICATION_CENTER_MEMBERSHIPS_TABLE_NAME = 'certification-center-memberships';
